pyproject.toml
[project]
name = "aqwel-aion"
version = "0.1.9"
description = "Complete AI Research & Development Library with Advanced Mathematics and ML Tools"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Aksel Aghajanyan"}
]
maintainers = [
    {name = "Aqwel AI Team"}
]

keywords = ["ai-research", "machine-learning", "mathematics", "statistics", "data-science", "embeddings", "aqwel-ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "watchdog>=2.1.0",
    "gitpython>=3.1.0",
]

[project.urls]
Homepage = "https://aqwelai.xyz/"
Documentation = "https://aqwelai.xyz/#/docs"
PyPI = "https://pypi.org/project/aqwel-aion/"

[project.optional-dependencies]
tools = [
    "tiktoken>=0.5.0",
]
rag = [
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
]
config = [
    "tomli>=2.0.0; python_version < \"3.11\"",
    "pyyaml>=6.0",
]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
]
viz = [
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]
former = [
    "matplotlib>=3.5.0",
    "pyyaml>=6.0",
]
ai = [
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "transformers>=4.20.0",
    "torch>=1.12.0",
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
]
docs = [
    "reportlab>=3.6.0",
    "pillow>=9.0.0",
]
full = [
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "transformers>=4.20.0",
    "torch>=1.12.0",
    "openai>=1.0.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "reportlab>=3.6.0",
    "pillow>=9.0.0",
    "tiktoken>=0.5.0",
    "tomli>=2.0.0; python_version < \"3.11\"",
    "pyyaml>=6.0",
]
metrics = [
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel", "pybind11>=2.10.0"]
build-backend = "setuptools.build_meta"